test_project/pyproject.toml
[tool.poetry]
name = "test-project"
version = "0.1.0"
description = ""
authors = ["sergkozun <kolya@mail.ru>"]
readme = "README.md"
#packages = [{include = "test_project"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.97.0"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.16"
alembic = "^1.11.1"
python-jose = "^3.3.0"
redis = "^4.5.5"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
pytest = "^7.4.0"
httpx = "^0.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
